% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw-seamove.R
\name{comp.movement.prob.by.age.season}
\alias{comp.movement.prob.by.age.season}
\title{Computing four seasonal vectors of movement probabilities by age for ri-->rj directly from SEAPODYM output files [spname]_FluxesRegion_age[a].txt. See also function get.movement.prob.by.age.season(), which reads the movement probabilities previously exported to the Multifan-CL format.}
\usage{
comp.movement.prob.by.age.season(dir, ri, rj, age = 0:50, print.out = FALSE)
}
\arguments{
\item{dir}{is the SEAPODYM output directory containing files [spname]_FluxesRegion_age[a].txt, where a is the age class indices from 0 to A+.}

\item{ri}{and rj are two regions, so that the vector of probabilities at age will be derived for ri->rj movement.}

\item{age}{is the vector of selecled age class indices, must be a sub-set of age classes written in the outputs.}

\item{print.out}{is the logical parameter, can be useful in case of large files to follow progression.}
}
\value{
Returns the vector of probabilities of movement from region ri to rj for the age classes 'age'.
}
\description{
Computing four seasonal vectors of movement probabilities by age for ri-->rj directly from SEAPODYM output files [spname]_FluxesRegion_age[a].txt. See also function get.movement.prob.by.age.season(), which reads the movement probabilities previously exported to the Multifan-CL format.
}
\examples{
i<-1; j<-2; age<-0:47
move_ij<-comp.movement.prob.by.age.season(dir.out,i,j,age)
}
