% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw-seamove.R
\name{get.movement.prob.by.age.season}
\alias{get.movement.prob.by.age.season}
\title{Function reads the Multifan-CL file with movement probabilities with seasonal and age structure.}
\usage{
get.movement.prob.by.age.season(filename, ri, rj, nbr, age)
}
\arguments{
\item{filename}{is the name of the Multifan-CL file with movement probabilities.}

\item{ri}{and rj are two regions, so that the vector of probabilities at age describe ri->rj movement.}

\item{age}{is the vector of Multifan-CL age class indices, beginning with 1.}
}
\value{
Returns the vector of probabilities of movement from region ri to rj for the age classes 'age'.
}
\description{
Function reads the Multifan-CL file with movement probabilities with seasonal and age structure.
}
\examples{
i<-1; j<-2; age<-0:47
move_ij<-get.movement.prob.by.age.season(mfcl.file,i,j,age)
}
