% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-seamove.R
\name{plot.fluxes.region}
\alias{plot.fluxes.region}
\title{Plotting SEAPODYM fluxes for a given region in PDF file.}
\usage{
\method{plot}{fluxes.region}(dir, reg, age = 0:18, year = NULL, treso = "quarterly", add.B = FALSE)
}
\arguments{
\item{dir}{is the SEAPODYM output directory containing files [spname]_FluxesRegion_age[a].txt, where a is the age class indices from 0 to A+.}

\item{reg}{is the region for which the fluxes, i.e., biomass flow per deltaT from and to 'reg', will be plotted.}

\item{age}{is the vector of age class indices, selected for plotting. Must be a sub-set of age classes written in the outputs.}

\item{year}{by default is set to NULL, but it can be set to a value within the temporal range in the outputs.}

\item{treso}{is the temporal resolution of fluxes that will be plotted, can be either "monthly" or "quarterly". If the outputs are written as quarterly, both options will provide the same quarterly visualization. If outputs are monthly and treso is quarterly, then the monthly fluxes will be aggregated into quarterly.}

\item{add.B}{is logical parameter, if set to TRUE, then the total biomass in the region 'reg' before the movement will be added to the plot on the right-hand y axis.}
}
\description{
The monthly or quatrerly mean values averaged over the time series written in the output files, or the monthly or quarterly fluxes for a selected year. This function uses global environment variables 'sp' and 'figs.dir'. If 'sp' was not set, it will be reset from the file names, which are conventional. Same for figs.dir, it will be created in the working directory if it was not defined earlier.
}
\examples{
plot.fluxes.region("./output/",1,3:10) # plot fluxes from and into region 1 for only immature adults (if skipjack)
}
