% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-seamove.R
\name{plot.movement.prob.by.age.season}
\alias{plot.movement.prob.by.age.season}
\title{Plotting seasonal movement probabilities by age between two regions, either read from 'movement-matrices-for-MFCL.txt' file or computed from SEAPODYM outputs.}
\usage{
\method{plot}{movement.prob.by.age.season}(move_ij, ri, rj)
}
\arguments{
\item{move_ij}{is the list containing the age vector and the seasonal movement probabilities at age from region 'i' to region 'j', stored as a 2d array with dimensions 4 (seasons) and the number of age classes. The age structure can be as defined in the simulation, or aggregated into quarterly classes, which is typical resolution of Multifan-CL.}

\item{ri}{and rj are regional indices, which should be in the range from 1 to 'nreg', with 'nreg' being the total number of regions in the regional structure defined in seapodym_fluxes simulations.}
}
\description{
Plotting seasonal movement probabilities by age between two regions, either read from 'movement-matrices-for-MFCL.txt' file or computed from SEAPODYM outputs.
}
\examples{
Example with MFCL file
file.name<-"./movement-matrices-for-MFCL.txt"
i<-1; j<-2; 
move_ij<-get.movement.prob.by.age.season(file.name,i,j,nbr,nba)	
plot.movement.prob.by.age.season(move_ij,i,j)
}
