% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw-seamove.R
\name{read.data.fluxes.header}
\alias{read.data.fluxes.header}
\title{Reading the header of the SEAPODYM output file with movement matrices.}
\usage{
read.data.fluxes.header(filename)
}
\arguments{
\item{Filename}{is the full name (with path) of the file to read.}
}
\value{
Returns the list with species name, mean age and length, as well as number of regions and dates written in the 'fluxes' files.
}
\description{
Reading the header of the SEAPODYM output file with movement matrices.
}
