% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw-seamove.R
\name{read.data.fluxes}
\alias{read.data.fluxes}
\title{Reading the SEAPODYM output file with movement matrices.}
\usage{
read.data.fluxes(filename, years = NULL)
}
\arguments{
\item{filename}{is the full name (with path) of the file to read.}

\item{years}{can be given in case the movement matrices should be extracted for a given sub-set of years. The default value is NULL.}
}
\value{
Returns the matrices with SEAPODYM movement fluxes.
}
\description{
Reading the SEAPODYM output file with movement matrices.
}
\examples{
filename<-"output/skj_FluxesRegion_age31.txt"
dat<-read.data.fluxes(filename,2001:2010)
}
