% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw-seamove.R
\name{sea2mfcl.movement}
\alias{sea2mfcl.movement}
\title{Converting the SEAPODYM movement fluxes to the Multifan-CL movement probabilities.}
\usage{
sea2mfcl.movement(
  dir.in,
  dir.out,
  years = NULL,
  age.in,
  age.plus = TRUE,
  aggregate.age = 3,
  nbr.in,
  aggregate.reg = FALSE,
  reg.mfcl = NULL,
  plot.out = TRUE
)
}
\arguments{
\item{dir.in}{is the SEAPODYM output directory containing files [spname]_FluxesRegion_age[a].txt, where a is the age class indices from 0 to A+. It can either be the original output with the original regional and age structures defined in simulations, or the aggregated outputs.}

\item{dir.out}{is the new output directory containing files [spname]_FluxesRegion_age[a].txt with the regional and age structures defined in Multifan-CL. The file with movement probabilities will be written in this directory as well.}

\item{years}{can be given in case the movement matrices should be extracted for a given sub-set of years. The default value is NULL.}

\item{age.in}{is the vector of age class indices, determines the full or a sub-set of age classes to be extracted from the outputs written in dir.in.}

\item{age.plus}{has the logical value, indicating whether the last age class in age.in is the A+. If yes, it will be used as is in the aggregation to Multifan-CL structure. Otherwise, it will be included in the aggregation.}

\item{aggregate.age}{is the size of the Multifan-CL age class in SEAPODYM age class units. If '0', then no aggregation will be done.}

\item{nbr.in}{is the number of regions in the SEAPODYM outputs written in dir.in.}

\item{aggregate.reg}{has the logical value indicating whether the SEAPODYM regional structure should be mapped to Multifan-CL. See reg.mfcl and function reg.structure.mapping() for more details.}

\item{reg.mfcl}{is the list with MFCL regions, where each element of the list contains the indices of the seapodym regions. Default value is NULL.}

\item{plot.out}{activates plotting of resulting movement probabilities for Multifan-CL regions and age structure. Defalt value is TRUE. Note, the PNG files with plots will be written in figs.dir in the working directory.}
}
\description{
Converting the SEAPODYM movement fluxes to the Multifan-CL movement probabilities.
}
\examples{
# Example #1: skipjack model with 48 monthly age classes and 15 regions, 
# extracting movement rates for age classes older than 3 months of age,
# i.e., skipping first three monthly classes, the 48th being the A+:
# First, define the parent SKJ.DIR, where the outputs are stored.
dir.in<-paste0(SKJ.DIR,"/output/")
dir.out<-paste0(SKJ.DIR,"/output-mfcl/")
# Second, provide the regional mapping between SEAPODYM and Multifan-CL regions
reg.mfcl<-list(r1=c(1,2),r2=c(3,4),r3=c(5,6),r4=7,r5=c(8,9),r6=c(10,11),r7=c(12,13,14),r8=15)
# Third, call the function:
sea2mfcl.movement(dir.in,dir.out,1983:2010,age.in=2:47,age.plus=TRUE,aggregate.age=3,nbr.in=15,aggregate.reg=TRUE,reg.mfcl)

#Example #2: albacore model configuration
dir.in<-paste0(ALB.DIR,"/output/")
dir.out<-paste0(ALB.DIR,"/output-mfcl/")
figs.dir<-paste0(ALB.DIR,"figs/figs-libtest/")
sea2mfcl.movement(dir.in,dir.out,1979:2010,age.in=5:148,age.plus=FALSE,aggregate.age=3,nbr.in=4,aggregate.reg=FALSE) 
}
